Add tests and update coments
